+++ Ruddarr/Dependencies/API/API+Live.swift
@@ -2,57 +2,63 @@ import Foundation
 
 extension API {
     static var live: Self {
-        .init(fetchMovies: { instance in
+        .init(radarr: .live, sonarr: .live, instance: .live)
+    }
+}
+
+extension RadarrAPI {
+    static var live: Self {
+        .init(fetch: { instance in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/movie")
 
-            var movies: [Movie] = try await request(url: url, instance: instance, timeout: .slow)
+            var movies: [Movie] = try await API.request(url: url, instance: instance, timeout: .slow)
             movies.stamp(instance.id)
             return movies
-        }, lookupMovies: { instance, query in
+        }, lookup: { instance, query in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/movie/lookup")
                 .appending(queryItems: [.init(name: "term", value: query)])
 
-            return try await request(url: url, instance: instance, timeout: .sluggish)
-        }, lookupMovieReleases: { movieId, instance in
+            return try await API.request(url: url, instance: instance, timeout: .sluggish)
+        }, releases: { movieId, instance in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/release")
                 .appending(queryItems: [.init(name: "movieId", value: String(movieId))])
 
-            return try await request(url: url, instance: instance, timeout: .releaseSearch)
-        }, getMovie: { movieId, instance in
+            return try await API.request(url: url, instance: instance, timeout: .releaseSearch)
+        }, movie: { movieId, instance in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/movie")
                 .appending(path: String(movieId))
 
-            var movie: Movie = try await request(url: url, instance: instance)
+            var movie: Movie = try await API.request(url: url, instance: instance)
             movie.stamp(instance.id)
             return movie
-        }, getMovieHistory: { movieId, instance in
+        }, history: { movieId, instance in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/history/movie")
                 .appending(queryItems: [.init(name: "movieId", value: String(movieId))])
 
-            return try await request(url: url, instance: instance)
-        }, getMovieFiles: { movieId, instance in
+            return try await API.request(url: url, instance: instance)
+        }, files: { movieId, instance in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/moviefile")
                 .appending(queryItems: [.init(name: "movieId", value: String(movieId))])
 
-            return try await request(url: url, instance: instance)
-        }, getMovieExtraFiles: { movieId, instance in
+            return try await API.request(url: url, instance: instance)
+        }, extraFiles: { movieId, instance in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/extrafile")
                 .appending(queryItems: [.init(name: "movieId", value: String(movieId))])
 
-            return try await request(url: url, instance: instance)
-        }, addMovie: { movie, instance in
+            return try await API.request(url: url, instance: instance)
+        }, add: { movie, instance in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/movie")
 
-            return try await request(method: .post, url: url, body: movie, instance: instance)
-        }, updateMovie: { movie, moveFiles, instance in
+            return try await API.request(method: .post, url: url, body: movie, instance: instance)
+        }, update: { movie, moveFiles, instance in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/movie/editor")
 
@@ -67,8 +73,8 @@ extension API {
                 moveFiles: moveFiles ? true : nil
             )
 
-            return try await request(method: .put, url: url, body: body, instance: instance)
-        }, deleteMovie: { movie, addExclusion, deleteFildes, instance in
+            return try await API.request(method: .put, url: url, body: body, instance: instance)
+        }, delete: { movie, addExclusion, deleteFildes, instance in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/movie")
                 .appending(path: String(movie.id))
@@ -77,21 +83,39 @@ extension API {
                     .init(name: "addImportExclusion", value: addExclusion ? "true" : "false"),
                 ])
 
-            return try await request(method: .delete, url: url, instance: instance)
-        }, deleteMovieFile: { file, instance in
+            return try await API.request(method: .delete, url: url, instance: instance)
+        }, deleteFile: { file, instance in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/moviefile")
                 .appending(path: String(file.id))
 
-            return try await request(method: .delete, url: url, instance: instance)
-        }, fetchSeries: { instance in
+            return try await API.request(method: .delete, url: url, instance: instance)
+        }, calendar: { start, end, instance in
+            let url = try await instance.baseURL()
+                .appending(path: "/api/v3/calendar")
+                .appending(queryItems: [
+                    .init(name: "unmonitored", value: "true"),
+                    .init(name: "start", value: start.formatted(.iso8601)),
+                    .init(name: "end", value: end.formatted(.iso8601)),
+                ])
+
+            var movies: [Movie] = try await API.request(url: url, instance: instance, timeout: .slow)
+            movies.stamp(instance.id)
+            return movies
+        })
+    }
+}
+
+extension SonarrAPI {
+    static var live: Self {
+        .init(fetch: { instance in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/series")
 
-            var series: [Series] = try await request(url: url, instance: instance, timeout: .slow)
+            var series: [Series] = try await API.request(url: url, instance: instance, timeout: .slow)
             series.stamp(instance.id)
             return series
-        }, fetchEpisodes: { seriesId, instance in
+        }, episodes: { seriesId, instance in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/episode")
                 .appending(queryItems: [
@@ -99,16 +123,16 @@ extension API {
                     .init(name: "includeEpisodeFile", value: "true"),
                 ])
 
-            var episodes: [Episode] = try await request(url: url, instance: instance)
+            var episodes: [Episode] = try await API.request(url: url, instance: instance)
             episodes.stamp(instance.id)
             return episodes
-        }, lookupSeries: { instance, query in
+        }, lookup: { instance, query in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/series/lookup")
                 .appending(queryItems: [.init(name: "term", value: query)])
 
-            return try await request(url: url, instance: instance, timeout: .sluggish)
-        }, lookupSeriesReleases: { seriesId, seasonId, episodeId, instance in
+            return try await API.request(url: url, instance: instance, timeout: .sluggish)
+        }, releases: { seriesId, seasonId, episodeId, instance in
             var url = try await instance.baseURL()
                 .appending(path: "/api/v3/release")
 
@@ -118,27 +142,27 @@ extension API {
                 url = url.appending(queryItems: [.init(name: "seriesId", value: String(seriesId!)), .init(name: "seasonNumber", value: String(seasonId!))])
             }
 
-            return try await request(url: url, instance: instance, timeout: .releaseSearch)
-        }, getSeries: { seriesId, instance in
+            return try await API.request(url: url, instance: instance, timeout: .releaseSearch)
+        }, series: { seriesId, instance in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/series")
                 .appending(path: String(seriesId))
 
-            var series: Series = try await request(url: url, instance: instance)
+            var series: Series = try await API.request(url: url, instance: instance)
             series.stamp(instance.id)
             return series
-        }, addSeries: { series, instance in
+        }, add: { series, instance in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/series")
 
-            return try await request(method: .post, url: url, body: series, instance: instance)
-        }, pushSeries: { series, instance in
+            return try await API.request(method: .post, url: url, body: series, instance: instance)
+        }, push: { series, instance in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/series")
                 .appending(path: String(series.id))
 
-            return try await request(method: .put, url: url, body: series, instance: instance)
-        }, updateSeries: { series, moveFiles, instance in
+            return try await API.request(method: .put, url: url, body: series, instance: instance)
+        }, update: { series, moveFiles, instance in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/series/editor")
 
@@ -155,8 +179,8 @@ extension API {
                 moveFiles: moveFiles ? true : nil
             )
 
-            return try await request(method: .put, url: url, body: body, instance: instance)
-        }, deleteSeries: { series, addExclusion, deleteFiles, instance in
+            return try await API.request(method: .put, url: url, body: body, instance: instance)
+        }, delete: { series, addExclusion, deleteFiles, instance in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/series")
                 .appending(path: String(series.id))
@@ -165,46 +189,34 @@ extension API {
                     .init(name: "addImportListExclusion", value: addExclusion ? "true" : "false"),
                 ])
 
-            return try await request(method: .delete, url: url, instance: instance)
+            return try await API.request(method: .delete, url: url, instance: instance)
         }, monitorEpisode: { ids, monitored, instance in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/episode/monitor")
 
             let body = EpisodesMonitorResource(episodeIds: ids, monitored: monitored)
 
-            return try await request(method: .put, url: url, body: body, instance: instance)
-        }, getEpisodeHistory: { id, instance in
+            return try await API.request(method: .put, url: url, body: body, instance: instance)
+        }, episodeHistory: { id, instance in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/history")
                 .appending(queryItems: [.init(name: "episodeId", value: String(id))])
 
-            return try await request(url: url, instance: instance)
+            return try await API.request(url: url, instance: instance)
         }, deleteEpisodeFile: { file, instance in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/episodefile")
                 .appending(path: String(file.id))
 
-            return try await request(method: .delete, url: url, instance: instance)
+            return try await API.request(method: .delete, url: url, instance: instance)
         }, deleteEpisodeFiles: { files, instance in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/episodefile/bulk")
 
             let body = EpisodeDeleteResource(episodeFileIds: files.map(\.id))
 
-            return try await request(method: .delete, url: url, body: body, instance: instance)
-        }, movieCalendar: { start, end, instance in
-            let url = try await instance.baseURL()
-                .appending(path: "/api/v3/calendar")
-                .appending(queryItems: [
-                    .init(name: "unmonitored", value: "true"),
-                    .init(name: "start", value: start.formatted(.iso8601)),
-                    .init(name: "end", value: end.formatted(.iso8601)),
-                ])
-
-            var movies: [Movie] = try await request(url: url, instance: instance, timeout: .slow)
-            movies.stamp(instance.id)
-            return movies
-        }, episodeCalendar: { start, end, instance in
+            return try await API.request(method: .delete, url: url, body: body, instance: instance)
+        }, calendar: { start, end, instance in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/calendar")
                 .appending(queryItems: [
@@ -214,44 +226,50 @@ extension API {
                     .init(name: "end", value: end.formatted(.iso8601)),
                 ])
 
-            var episodes: [Episode] = try await request(url: url, instance: instance, timeout: .slow)
+            var episodes: [Episode] = try await API.request(url: url, instance: instance, timeout: .slow)
             episodes.stamp(instance.id)
             return episodes
-        }, command: { command, instance in
+        })
+    }
+}
+
+extension InstanceAPI {
+    static var live: Self {
+        .init(command: { command, instance in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/command")
 
-            return try await request(method: .post, url: url, body: command.payload, instance: instance)
+            return try await API.request(method: .post, url: url, body: command.payload, instance: instance)
         }, downloadRelease: { payload, instance in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/release")
 
-            return try await request(method: .post, url: url, body: payload, instance: instance, timeout: .sluggish)
-        }, systemStatus: { instance in
+            return try await API.request(method: .post, url: url, body: payload, instance: instance, timeout: .sluggish)
+        }, status: { instance in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/system/status")
 
-            return try await request(url: url, instance: instance)
+            return try await API.request(url: url, instance: instance)
         }, rootFolders: { instance in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/rootfolder")
 
-            return try await request(url: url, instance: instance, timeout: .slow)
+            return try await API.request(url: url, instance: instance, timeout: .slow)
         }, qualityProfiles: { instance in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/qualityprofile")
-            return try await request(url: url, instance: instance)
-        }, fetchDiskSpace: { instance in
+            return try await API.request(url: url, instance: instance)
+        }, diskSpace: { instance in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/diskspace")
 
-            return try await request(url: url, instance: instance)
-        }, getTags: { instance in
+            return try await API.request(url: url, instance: instance)
+        }, tags: { instance in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/tag")
 
-            return try await request(url: url, instance: instance)
-        }, fetchQueueTasks: { instance in
+            return try await API.request(url: url, instance: instance)
+        }, queue: { instance in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/queue")
                 .appending(queryItems: [
@@ -261,7 +279,7 @@ extension API {
                     .init(name: "pageSize", value: "250"),
                 ])
 
-            var items: QueueItems = try await request(url: url, instance: instance)
+            var items: QueueItems = try await API.request(url: url, instance: instance)
             items.records.stamp(instance.id)
             return items
         }, deleteQueueTask: { task, remove, block, search, instance in
@@ -274,8 +292,8 @@ extension API {
                     .init(name: "skipRedownload", value: search ? "false" : "true"),
                 ])
 
-            return try await request(method: .delete, url: url, instance: instance)
-        }, fetchImportableFiles: { downloadId, instance in
+            return try await API.request(method: .delete, url: url, instance: instance)
+        }, importableFiles: { downloadId, instance in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/manualimport")
                 .appending(queryItems: [
@@ -283,8 +301,8 @@ extension API {
                     .init(name: "filterExistingFiles", value: "false"),
                 ])
 
-            return try await request(url: url, instance: instance, timeout: .sluggish)
-        }, fetchHistory: { type, page, limit, instance in
+            return try await API.request(url: url, instance: instance, timeout: .sluggish)
+        }, history: { type, page, limit, instance in
             var url = try await instance.baseURL()
                 .appending(path: "/api/v3/history")
                 .appending(queryItems: [
@@ -296,31 +314,31 @@ extension API {
                 url = url.appending(queryItems: [.init(name: "eventType", value: String(type))])
             }
 
-            var history: MediaHistory = try await request(url: url, instance: instance)
+            var history: MediaHistory = try await API.request(url: url, instance: instance)
             history.records.stamp(instance.id)
             return history
-        }, fetchNotifications: { instance in
+        }, notifications: { instance in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/notification")
 
-            return try await request(url: url, instance: instance)
+            return try await API.request(url: url, instance: instance)
         }, createNotification: { model, instance in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/notification")
 
-            return try await request(method: .post, url: url, body: model, instance: instance)
+            return try await API.request(method: .post, url: url, body: model, instance: instance)
         }, updateNotification: { model, instance in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/notification")
                 .appending(path: String(model.id ?? 0))
 
-            return try await request(method: .put, url: url, body: model, instance: instance)
+            return try await API.request(method: .put, url: url, body: model, instance: instance)
         }, deleteNotification: { model, instance in
             let url = try await instance.baseURL()
                 .appending(path: "/api/v3/notification")
                 .appending(path: String(model.id ?? 0))
 
-            return try await request(method: .delete, url: url, instance: instance)
+            return try await API.request(method: .delete, url: url, instance: instance)
         })
     }
 }