+++ Ruddarr/Models/Movies/MovieMetadata.swift
@@ -43,8 +43,8 @@ class MovieMetadata {
         filesError = false
 
         do {
-            async let filesResult = dependencies.api.getMovieFiles(movie.id, instance)
-            async let extraFilesResult = dependencies.api.getMovieExtraFiles(movie.id, instance)
+            async let filesResult = dependencies.api.radarr.files(movie.id, instance)
+            async let extraFilesResult = dependencies.api.radarr.extraFiles(movie.id, instance)
 
             files = try await filesResult
             extraFiles = try await extraFilesResult
@@ -66,7 +66,7 @@ class MovieMetadata {
         historyError = false
 
         do {
-            history = try await dependencies.api.getMovieHistory(movie.id, instance)
+            history = try await dependencies.api.radarr.history(movie.id, instance)
         } catch is CancellationError {
             // do nothing
         } catch {
@@ -77,9 +77,12 @@ class MovieMetadata {
     }
 
     func refresh(for movie: Movie) async {
+        filesError = false
+        historyError = false
+
         do {
-            async let filesResult = dependencies.api.getMovieFiles(movie.id, instance)
-            async let extraFilesResult = dependencies.api.getMovieExtraFiles(movie.id, instance)
+            async let filesResult = dependencies.api.radarr.files(movie.id, instance)
+            async let extraFilesResult = dependencies.api.radarr.extraFiles(movie.id, instance)
 
             files = try await filesResult
             extraFiles = try await extraFilesResult
@@ -92,7 +95,7 @@ class MovieMetadata {
         filesLoading = false
 
         do {
-            history = try await dependencies.api.getMovieHistory(movie.id, instance)
+            history = try await dependencies.api.radarr.history(movie.id, instance)
         } catch is CancellationError {
             // do nothing
         } catch {
@@ -104,7 +107,7 @@ class MovieMetadata {
 
     func delete(_ file: MediaFile) async -> Bool {
         do {
-            _ = try await dependencies.api.deleteMovieFile(file, instance)
+            _ = try await dependencies.api.radarr.deleteFile(file, instance)
 
             if let index = files.firstIndex(where: { $0.id == file.id }) {
                 files.remove(at: index)