+++ Ruddarr/Views/Activity/TaskRemovalView.swift
@@ -86,7 +86,7 @@ struct TaskRemovalView: View {
         }
 
         do {
-            _ = try await dependencies.api.deleteQueueTask(
+            _ = try await dependencies.api.instance.deleteQueueTask(
                 item.id, remove, block, search, instance
             )
         } catch is CancellationError {
@@ -100,6 +100,8 @@ struct TaskRemovalView: View {
         }
 
         if error == nil {
+            Queue.shared.markDeleted(item)
+
             await Queue.shared.fetchTasks()
         }