+++ Ruddarr/Views/ContentView.swift
@@ -53,6 +53,7 @@ struct ContentView: View {
         .displayToasts()
         .whatsNewSheet()
         .reportBugSheet()
+        // .testFlightNotice()
     }
 
     var movies: TabItem { .movies }