+++ Ruddarr/Views/SeriesView.swift
@@ -365,7 +365,7 @@ struct SeriesView: View {
 }
 
 #Preview("Offline") {
-    dependencies.api.fetchSeries = { _ in
+    dependencies.api.sonarr.fetch = { _ in
         throw API.Error.notConnectedToInternet
     }