+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to Ruddarr are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## 2.0.1 - 2026-08-19
+### Changed
+- Deleted tasks remaining in the queue until the next refresh
+- Collapse descriptions when tapping again
+
+### Fixed
+- Fixed a false connection failure when during cold start
+- Fixed a stale movie files being shown after a successful refresh
+
 ## 2.0.0 - 2026-08-12
 ### Added
 - Support alternate instance URLs and automatic network detection